requirements.txt
langchain                     ==1.2.0
langchain-core                ==1.2.6
langchain-community           ==0.4.1
langchain-groq                ==1.1.1
langchain-text-splitters      ==1.1.0
streamlit                     ==1.52.2
python-dotenv                 ==1.2.1
setuptools                    ==80.9.0
pillow                        ==12.1.0
langchain-tavily
langchain-openai
langchain-chroma
langchain_huggingface
plotly
scikit-learn